feat(admin): views & producs control
- Añadidas vistas para las opciones de gestion de los recursos de la app, desde la pestaña de administrador
- Implementada funcion para la carga masiva de prodcutos mediante archivo excel
- Implementada paginacion de la tabla de productos para conseguir mayor eficiencia en la carga de la vista

+++ src/main/java/es/laboticademar/webstore/controllers/ProductController.java
@@ -1,18 +1,23 @@
 package es.laboticademar.webstore.controllers;
 
-import java.util.List;
+import java.math.BigDecimal;
+import java.util.HashMap;
+import java.util.Map;
 
+import org.springframework.data.domain.Page;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.data.domain.PageRequest;
+import org.springframework.data.domain.Sort;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RequestParam;
 
-import es.laboticademar.webstore.entities.Usuario;
-import es.laboticademar.webstore.services.DestacadoService;
-import es.laboticademar.webstore.services.ProductService;
-import es.laboticademar.webstore.services.UsuarioService;
+import es.laboticademar.webstore.entities.Producto;
+import es.laboticademar.webstore.services.interfaces.DestacadoService;
+import es.laboticademar.webstore.services.interfaces.ProductService;
 
 
 @Controller
@@ -25,12 +30,12 @@ public class ProductController {
     private DestacadoService destacadoService;
 
     @GetMapping("/{id}")
-    public String goProduct(@PathVariable("id") Integer id, Model model) {
+    public String goProduct(@PathVariable("id") BigDecimal id, Model model) {
 
         model.addAttribute("destacados", destacadoService.getAllDestacados());
         model.addAttribute("producto", productService.getProductoById(id));
 
         return "product/product";
     }
-    
+
 }
