+++ src/main/java/es/laboticademar/webstore/repositories/ProductDAO.java
@@ -1,9 +1,8 @@
 package es.laboticademar.webstore.repositories;
 
+import java.math.BigDecimal;
 import org.springframework.data.jpa.repository.JpaRepository;
-
 import es.laboticademar.webstore.entities.Producto;
 
-public interface ProductDAO extends JpaRepository<Producto, Integer> {
-    
+public interface ProductDAO extends JpaRepository<Producto, BigDecimal> {
 }
